refactor(map): extract generation layer into lib/src/gen/ (#3574) (#3586)
* refactor(map): extract generation layer into lib/src/gen/ (#3574)

Completes the explicit gen/view/render internal layering for
colonizethis_map (slice 5, generation portion). The view/ and render/
layers were relocated in prior slices (#3584, #3585); this moves the
deterministic tile-map generation pipeline and its generation-only
helpers into packages/colonizethis_map/lib/src/gen/, leaving cross-layer
primitives (grid traversal, cardinal directions, region dispatch,
topology helpers, tile-key/pipe utils, validation) as shared modules at
the lib/src/ root.

Pure relocation: no behavior change. Inter-generation imports stay
same-dir; imports of shared-root modules become `../`. The public
colonizethis_map.dart barrel re-exports the relocated modules, so
consumers (app/, ctdev/, tool/) are source-compatible.

- Move 31 generation files (tile_map_generator + parts, land-seeds +
  parts, voronoi, topology gen/inference, distance transforms,
  MapGenStage/MapGenPass contracts, resource placement/cap state,
  grid graph, locked-pair fixture) into lib/src/gen/.
- Update barrel exports, DI provider import, and same-package test
  deep-imports to src/gen/.
- Update repo-lint scripts + fixtures keyed on generation paths
  (map_gen_stage_protocol, map_no_partfile_classes,
  map_grid_cell_iteration_central exemptions, public_barrel_surface
  forbidden modules) and the gen-no-image-import doc to reference gen/.
- SPEC: document the generation layer folder in map-visualization.md
  § Internal layering and refresh repo-lint.md path references.

Verified: dart analyze (map lib clean), 296 map package tests pass,
all 124 ct_repo_lint rules pass, affected repo-lint unit tests green.

Refs #3574


* refactor(map): collapse join-sea part fragments into one file (#3574)

Slice 7: consolidate the three sibling join-sea part fragments
(bridge / jitter / subdivide) into tile_map_generator_join_sea.dart as
cohesive extensions on _TileMapGenJoinSea, reducing generator part-file
sprawl. Each fragment held a single private top-level extension and the
parent join-sea file stays well under the 1000-line part-fragment cap
(563 lines). Pure relocation; no behavior change.

Refs #3574


* refactor(map): relocate merged hardwood-cluster part into lib/src/gen/ (#3574)

dev (#3573/#3575) added tile_map_generator_terrain_hardwood_part.dart at
the pre-relocation lib/src/ path; this branch had moved the generation
library to lib/src/gen/. Move the merged part file into gen/ so its
`part of 'tile_map_generator.dart'` resolves and the generator library's
part directive points at an existing file.

Refs #3574


---------
