+++ packages/colonizethis_map/lib/src/gen/map_partition_gates_exhausted.dart
