+++ packages/colonizethis_map/lib/src/gen/tile_map_distance_sentinels.dart
