+++ packages/colonizethis_map/lib/src/gen/tile_map_generator_join_sea.dart
@@ -1,4 +1,25 @@
+/// Pass 10–11: join continents, terrain jitter, sea subdivision.
+///
+/// The continent-joining, terrain-jitter, and sea-subdivision passes were
+/// previously split across three sibling part fragments; they are consolidated
+/// here as cohesive extensions on [_TileMapGenJoinSea] (Refs #3574, slice 7).
+
 part of 'tile_map_generator.dart';
+
+/// Exempt from the uniform [MapGenPass] entry point (Refs #3574, slice 4):
+/// this family owns three heterogeneous passes — continent joining
+/// (grid+terrain+resource in/out), terrain jitter (in-place mutation), and
+/// sea-zone subdivision (grid in/out) — that share no single representative
+/// payload/result shape. It therefore implements [MapGenStage] only; the
+/// orchestrator drives its three passes via their dedicated methods.
+class _TileMapGenJoinSea implements MapGenStage {
+  _TileMapGenJoinSea(this.params, this._log, this._graph);
+
+  @override
+  final TileMapParams params;
+  final CtLogger _log;
+  final TileMapGridGraph _graph;
+}
 
 extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
   /// Join step: for each continent with >1 land component, connect two by a shortest path of sea cells. Returns (grid, terrainGrid, resourceGrid, didJoin).
@@ -316,5 +337,228 @@ extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
       restoredToSea.add((x, y));
     }
     return restoredToSea;
+  }
+}
+
+extension _TileMapGenJoinSeaJitterPart on _TileMapGenJoinSea {
+  bool _jitterTileIsTerrainOrProvinceEdge(
+    int x,
+    int y,
+    String provinceId,
+    TerrainType dominant,
+    List<List<String>> grid,
+    List<List<TerrainType?>> terrainGrid,
+    int width,
+    int height,
+    List<(int dx, int dy)> directions4,
+  ) {
+    for (final (dx, dy) in directions4) {
+      final nx = x + dx;
+      final ny = y + dy;
+      if (nx < 0 || nx >= width || ny < 0 || ny >= height) continue;
+      final neighborProvince = grid[ny][nx];
+      final neighborTerrain = terrainGrid[ny][nx];
+      if (neighborProvince != provinceId) return true;
+      if (neighborTerrain != null && neighborTerrain != dominant) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  void jitterTerrainByProvince(
+    List<List<String>> grid,
+    List<List<TerrainType?>> terrainGrid,
+    List<List<Resource?>> resourceGrid,
+    String regionId,
+    Random rnd,
+  ) {
+    final allowedNonMountain = allowedTerrainsForRegion(
+      regionId,
+    ).where((t) => t != TerrainType.mountain).toList();
+    if (allowedNonMountain.isEmpty) return;
+
+    final height = grid.length;
+    if (height == 0) return;
+    final width = grid[0].length;
+    final tilesByProvince = <String, List<(int x, int y)>>{};
+    final provinceIdPattern = RegExp(r'^p\d+$');
+    TileMapGrid.forEachCell(grid, (y, x, id) {
+      if (!provinceIdPattern.hasMatch(id)) return;
+      tilesByProvince.putIfAbsent(id, () => []).add((x, y));
+    });
+    if (tilesByProvince.isEmpty) return;
+
+    const directions4 = kTileMapDirections4;
+    const directions8 = kTileMapDirections8;
+
+    for (final entry in tilesByProvince.entries) {
+      final tiles = entry.value;
+      if (tiles.length < params.jitterMinProvinceSize) continue;
+
+      final counts = <TerrainType, int>{};
+      var terrainTiles = 0;
+      for (final (x, y) in tiles) {
+        final t = terrainGrid[y][x];
+        if (t == null) continue;
+        counts[t] = (counts[t] ?? 0) + 1;
+        terrainTiles++;
+      }
+      if (terrainTiles == 0 || counts.isEmpty) continue;
+
+      TerrainType dominant = counts.keys.first;
+      var maxCount = counts[dominant]!;
+      for (final e in counts.entries) {
+        if (e.value > maxCount) {
+          dominant = e.key;
+          maxCount = e.value;
+        }
+      }
+      final fDom = maxCount / terrainTiles;
+      if (fDom < params.jitterHomogeneityThreshold) continue;
+
+      final candidates = <(int x, int y)>[];
+      for (final (x, y) in tiles) {
+        if (terrainGrid[y][x] != dominant) continue;
+        if (resourceGrid[y][x] != null) continue;
+        if (!_jitterTileIsTerrainOrProvinceEdge(
+          x,
+          y,
+          entry.key,
+          dominant,
+          grid,
+          terrainGrid,
+          width,
+          height,
+          directions4,
+        )) {
+          continue;
+        }
+        candidates.add((x, y));
+      }
+      if (candidates.isEmpty) continue;
+
+      candidates.shuffle(rnd);
+      final maxChanges = (params.jitterMaxFraction * tiles.length).floor();
+      if (maxChanges <= 0) continue;
+      var changes = 0;
+      for (final (x, y) in candidates) {
+        if (changes >= maxChanges) break;
+        if (rnd.nextDouble() > params.jitterProbability) continue;
+
+        final neighborCounts = <TerrainType, int>{};
+        for (final (dx, dy) in directions8) {
+          final nx = x + dx;
+          final ny = y + dy;
+          if (nx < 0 || nx >= width || ny < 0 || ny >= height) continue;
+          if (grid[ny][nx] != entry.key) continue;
+          final nt = terrainGrid[ny][nx];
+          if (nt == null || nt == dominant || nt == TerrainType.mountain) {
+            continue;
+          }
+          neighborCounts[nt] = (neighborCounts[nt] ?? 0) + 1;
+        }
+
+        final supported = neighborCounts.entries
+            .where((e) => e.value >= params.jitterNeighborSupportThreshold)
+            .map((e) => e.key)
+            .toList();
+        if (supported.isEmpty) continue;
+        terrainGrid[y][x] = supported[rnd.nextInt(supported.length)];
+        changes++;
+      }
+    }
+  }
+}
+
+extension _TileMapGenJoinSeaSubdividePart on _TileMapGenJoinSea {
+  /// Pass 11: Subdivide sea so each zone has at most maxSeaZoneFraction * totalSea tiles.
+  /// Returns (newGrid, total sea zone count).
+  (List<List<String>>, int) subdivideSeaZonesWithCap(
+    List<List<String>> grid,
+    String seaZoneId,
+    int totalSea,
+  ) {
+    final components = _graph.connectedComponentsOfSea(grid, seaZoneId);
+    if (components.isEmpty) return (grid, 0);
+    final sorted = List<Set<(int x, int y)>>.from(components)
+      ..sort((a, b) {
+        final (minYa, minXa) = _graph.minYx(a);
+        final (minYb, minXb) = _graph.minYx(b);
+        if (minYa != minYb) return minYa.compareTo(minYb);
+        return minXa.compareTo(minXb);
+      });
+    final g = snapshotGrid(grid);
+    final maxPerZone = (params.maxSeaZoneFraction * totalSea).floor();
+    var nextSeaZoneIndex = 1;
+    for (final component in sorted) {
+      final size = component.length;
+      if (maxPerZone <= 0 || size <= maxPerZone) {
+        final id = 's$nextSeaZoneIndex';
+        nextSeaZoneIndex++;
+        for (final (x, y) in component) {
+          g[y][x] = id;
+        }
+        continue;
+      }
+      final K = (size / maxPerZone).ceil().clamp(1, size);
+      final seeds = _placeSeaSeedsFarthestPoint(component, K);
+      final seedMap = <String, (int x, int y)>{
+        for (var i = 0; i < seeds.length; i++)
+          's${nextSeaZoneIndex + i}': seeds[i],
+      };
+      final assignment = assignCellsToNearestSeed(
+        component,
+        seedMap,
+        noiseScale: params.voronoiNoiseScale,
+        noiseSeed: params.seed,
+      );
+      for (final entry in assignment.entries) {
+        final (x, y) = entry.key;
+        g[y][x] = entry.value;
+      }
+      nextSeaZoneIndex += K;
+    }
+    return (g, nextSeaZoneIndex - 1);
+  }
+
+  /// Place K well-spread seeds in [cells] using farthest-point sampling.
+  List<(int x, int y)> _placeSeaSeedsFarthestPoint(
+    Set<(int x, int y)> cells,
+    int K,
+  ) {
+    if (cells.isEmpty || K <= 0) return [];
+    final list = cells.toList();
+    if (K >= list.length) return list;
+    list.sort((a, b) {
+      if (a.$2 != b.$2) return a.$2.compareTo(b.$2);
+      return a.$1.compareTo(b.$1);
+    });
+    final chosen = <(int x, int y)>[list.first];
+    for (var i = 1; i < K; i++) {
+      chosen.add(_bestFarthestPointCellFromList(list, chosen));
+    }
+    return chosen;
+  }
+
+  (int x, int y) _bestFarthestPointCellFromList(
+    List<(int x, int y)> list,
+    List<(int x, int y)> chosen,
+  ) {
+    var bestCell = list.first;
+    var bestMinD2 = 0;
+    for (final (x, y) in list) {
+      if (chosen.contains((x, y))) continue;
+      var minD2 = kUnsetSquaredDistanceInt31;
+      for (final (sx, sy) in chosen) {
+        final d2 = (x - sx) * (x - sx) + (y - sy) * (y - sy);
+        if (d2 < minD2) minD2 = d2;
+      }
+      if (minD2 > bestMinD2) {
+        bestMinD2 = minD2;
+        bestCell = (x, y);
+      }
+    }
+    return bestCell;
   }
 }
