+++ packages/colonizethis_map/lib/src/gen/tile_map_generator_land_seeds.dart
@@ -15,10 +15,10 @@ import 'dart:math';
 import 'grid_voronoi.dart';
 import 'map_gen_pass_payloads.dart';
 import 'map_gen_stage.dart';
-import 'tile_map_directions.dart';
+import '../tile_map_directions.dart';
 import 'tile_map_distance_sentinels.dart';
 import 'tile_map_land_seed_contract.dart';
-import 'tile_map_grid.dart';
+import '../tile_map_grid.dart';
 import 'tile_map_land_sentinel.dart';
 import 'tile_map_manhattan_distance_maps.dart';
 import 'tile_map_manhattan_distance_transform.dart';
