+++ packages/colonizethis_map/lib/src/gen/tile_map_province_budget.dart
