+++ packages/colonizethis_map/lib/src/gen/tile_map_resource_cap_state.dart
