+++ packages/colonizethis_map/lib/src/gen/tile_map_resource_placement.dart
