+++ packages/colonizethis_map/test/tile_map_resource_cap_state_test.dart
@@ -1,5 +1,5 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_map/src/tile_map_resource_cap_state.dart';
+import 'package:colonizethis_map/src/gen/tile_map_resource_cap_state.dart';
 import 'package:colonizethis_test/test.dart';
 
 void main() {
