+++ packages/colonizethis_map/test/tile_map_resource_placement_test.dart
@@ -1,8 +1,8 @@
 import 'dart:math';
 
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_map/src/tile_map_resource_placement.dart';
-import 'package:colonizethis_map/src/tile_map_resource_cap_state.dart';
+import 'package:colonizethis_map/src/gen/tile_map_resource_placement.dart';
+import 'package:colonizethis_map/src/gen/tile_map_resource_cap_state.dart';
 import 'package:colonizethis_test/test.dart';
 
 void main() {
